extraction pipeline cleaning
